Evaluate from text API (#48)
Adding the /evaluate_from_file, which will take a text to evaluate, an OpenAI API key and a systematized concepts file (from #46), calls GPT-4o for each systematized concept and returns the results.

To implement that, a new module scorer.py is created with classes as functions to allow this evaluation, which are implementations of InspectAI Scorer abstractions.

Also moving the get_llm to its own file so we don't have weird imports.